Separate parsing logic from API Client logic
